Implemented getTaskCount-function
