+++ packages/task/src/internal.ts
@@ -3,5 +3,5 @@ export { _defaultHttpRequestFunctionFactory, _defaultHttpRequestFunction } from 
 export { _completeTaskFactory } from "./tasks/complete-task/complete-task";
 export { _createTaskFactory, _createTaskDefaultTransformFunction } from "./tasks/create-task/create-task";
 export { _deleteTaskFactory } from "./tasks/delete-task/delete-task";
-export { _getTaskFactory, _getTaskDefaultTransformFunction } from "./tasks/get-task/get-task";
+export { _getTaskCountFactory, _getTaskCountDefaultTransformFunction } from "./tasks/get-task-count/get-task-count";
 export { _updateTaskFactory } from "./tasks/update-task/update-task";
